local_data_process_service/pyproject.toml
[tool.poetry]
name = "local_data_process_service"
version = "0.1.0"
description = ""
authors = ["zun <trungdunglebui17112004@gmail.com>"]
readme = "README.md"
packages = [{include = "local_data_process_service", from = "src"}]


[tool.poetry.dependencies]
python = "^3.11"
adafruit-io = "2.8.0"
amqp = "5.3.1"
async-timeout = "5.0.1"
billiard = "4.2.1"
blinker = "1.9.0"
cachecontrol = "0.14.2"
cachetools = "5.5.2"
celery = "5.4.0"
certifi = "2025.1.31"
cffi = "1.17.1"
charset-normalizer = "3.4.1"
click = "8.1.8"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorlog = "6.9.0"
cryptography = "44.0.2"
dotenv = "0.9.9"
firebase-admin = "6.7.0"
flask = "3.1.0"
google-api-core = "2.24.2"
google-api-python-client = "2.165.0"
google-auth = "2.38.0"
google-auth-httplib2 = "0.2.0"
google-cloud-core = "2.4.3"
google-cloud-firestore = "2.20.1"
google-cloud-storage = "3.1.0"
google-crc32c = "1.7.0"
google-resumable-media = "2.7.2"
googleapis-common-protos = "1.69.2"
grpcio = "1.71.0"
grpcio-status = "1.71.0"
httplib2 = "0.22.0"
idna = "3.10"
influxdb-client = "1.48.0"
influxdb3-python = "0.11.0"
itsdangerous = "2.2.0"
jinja2 = "3.1.6"
kombu = "5.5.0"
markupsafe = "3.0.2"
msgpack = "1.1.0"
numpy = "2.2.4"
paho-mqtt = "2.1.0"
pandas = "2.2.3"
prompt-toolkit = "3.0.50"
proto-plus = "1.26.1"
protobuf = "5.29.4"
pyarrow = "19.0.1"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.1"
pycparser = "2.22"
pyjwt = "2.10.1"
pyparsing = "3.2.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2025.2"
reactivex = "4.0.4"
redis = "5.2.1"
requests = "2.32.3"
rsa = "4.9"
six = "1.17.0"
typing-extensions = "4.12.2"
tzdata = "2025.1"
uritemplate = "4.1.1"
urllib3 = "2.3.0"
vine = "5.1.0"
wcwidth = "0.2.13"
werkzeug = "3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"